Convert versionmanager to typescript.
