+++ src/index.ts
@@ -15,7 +15,7 @@ import _ from 'lodash'
 import getstdin from 'get-stdin'
 import Chalk from 'chalk'
 import jph from 'json-parse-helpfulerror'
-import * as vm from './versionmanager.js'
+import * as vm from './versionmanager'
 import doctor from './doctor.js'
 import * as packageManagers from './package-managers/index.js'
 import * as logging from './logging.js'
@@ -24,7 +24,7 @@ import cliOptions from './cli-options.js'
 import getNcurc from './lib/get-ncu-rc'
 import mergeOptions from './lib/mergeOptions'
 
-import { Index, Options } from './types'
+import { Index, Maybe, Options } from './types'
 
 const { deepPatternPrefix, doctorHelpText } = constants
 const { print, printJson, printUpgrades, printIgnoredUpdates } = logging
@@ -144,7 +144,7 @@ export function getPeerDependencies(current: Index<any>, options: Options) {
 }
 
 /** Checks local project dependencies for upgrades. */
-async function runLocal(options: Options, pkgData?: string | null, pkgFile?: string | null) {
+async function runLocal(options: Options, pkgData?: Maybe<string>, pkgFile?: Maybe<string>) {
 
   print(options, '\nOptions:', 'verbose')
   print(options, sortOptions(options), 'verbose')
@@ -193,7 +193,7 @@ async function runLocal(options: Options, pkgData?: string | null, pkgFile?: str
   print(options, '\nUpgraded:', 'verbose')
   print(options, upgraded, 'verbose')
 
-  const { newPkgData, selectedNewDependencies } = await vm.upgradePackageData(pkgData, current, upgraded, latest, options)
+  const { newPkgData, selectedNewDependencies } = await vm.upgradePackageData(pkgData!, current, upgraded, latest, options)
 
   const output = options.jsonAll ? jph.parse(newPkgData) :
     options.jsonDeps ?
@@ -232,7 +232,7 @@ async function runLocal(options: Options, pkgData?: string | null, pkgFile?: str
       ownersChangedDeps
     })
     if (options.peer) {
-      const ignoredUpdates = await vm.getIgnoredUpgrades(current, upgraded, upgradedPeerDependencies, options)
+      const ignoredUpdates = await vm.getIgnoredUpgrades(current, upgraded, upgradedPeerDependencies!, options)
       if (!_.isEmpty(ignoredUpdates)) {
         printIgnoredUpdates(options, ignoredUpdates)
       }
@@ -571,4 +571,4 @@ export async function run(options: Options = {}) {
 }
 
 export { default as getNcurc } from './lib/get-ncu-rc'
-export * from './versionmanager.js'
+export * from './versionmanager'
