+++ src/types/cint.d.ts
@@ -1,1 +1,5 @@
-declare module 'cint'
+declare module 'cint' {
+  export function filterObject<T = any>(obj: Index<T>, f: (key: string, value: T) => boolean): Index<T>
+  export function toArray<T = any, R = any>(object: Index<T>, f: (key: string, value: T) => R): R[]
+  export function toObject<T = any, R = any>(arr: T[], f: (value: T, i: number) => Index<R>): Index<R>
+}
